+++ src/data/pokemon/species_info/gen_1_families.h
